+++ poset.cc
@@ -6,9 +6,7 @@
 #include <vector>
 #include <cassert>
 
-
 #include "poset.h"
-
 
 #define assertm(exp, msg) assert(((void)msg, exp))
 
@@ -30,18 +28,20 @@ using std::to_string;
 using std::unordered_map;
 using std::vector;
 
-using index = size_t;
+using posetID_t = unsigned long;
+using idx_t = size_t;
 using poset_elem = string;
+// using inputVal_t = char const;
 using vectorOfStrings = vector<poset_elem>;
 using posetRelationsArray = vector<vector<int>>;
 using identificator = unsigned long;
 using poset_t = pair<vectorOfStrings *, posetRelationsArray *>;
 using allPosetsMap = unordered_map<identificator, poset_t *>;
 
+// unordered_map<identificator, poset_t *> allPosets;
 
 namespace
 {
-
   const int RELATION = 1;
   const int NO_RELATION = -1;
   const int RELATION_TRANSITIVITY = 2;
@@ -99,7 +99,7 @@ namespace
   // where elem_to_remove < elem2, than relation between elem1 < elem2
   // is removed.
   void deleteRelationsTransitivity(posetRelationsArray *relArr,
-                                   size_t currentElem, size_t idxOfElemToDelete, size_t nbrOfRows)
+                idx_t currentElem, idx_t idxOfElemToDelete, size_t nbrOfRows)
   {
     if (relArr->at(currentElem)[idxOfElemToDelete] == RELATION ||
         relArr->at(currentElem)[idxOfElemToDelete] == RELATION_TRANSITIVITY)
@@ -123,7 +123,7 @@ namespace
   // elem2 < elem_to_delete and if elem2 < elem1 transitivity
   // it deletes relation.
   void deleteRelationsLarger(posetRelationsArray *relArr,
-                             size_t currentElem, size_t idxOfElemToDelete, size_t nbrOfRows)
+                idx_t currentElem, idx_t idxOfElemToDelete, size_t nbrOfRows)
   {
     if (relArr->at(currentElem)[idxOfElemToDelete] == RELATION_IM_LARGER)
     {
@@ -140,9 +140,12 @@ namespace
   }
 
   void checkIfElemExistInVecOfStr(vectorOfStrings *v, char const *value,
-                                  size_t &idx, bool &exist)
+                                  idx_t &idx, bool &exist)
   {
+    idx = 0;
+    exist = false;
     size_t vSize = v->size();
+
     for (size_t i = 0; i < vSize; i++)
     {
       if ((*v)[i] == value)
@@ -154,37 +157,16 @@ namespace
     }
   }
 
-  // Function finds indexes of given values in vector that
-  // stores elements in poset. If element is not found its
-  // index is equal to -1.
-  void findIndexesOfGivenValues(long long &index1, long long &index2,
-                                char const *value1, char const *value2, vectorOfStrings *v)
-  {
-    index1 = NOT_FOUND;
-    index2 = NOT_FOUND;
-    size_t vSize = v->size();
-
-    for (index i = 0; i < vSize; i++)
-    {
-      if (v->at(i) == value1)
-        index1 = i;
-      if (v->at(i) == value2)
-        index2 = i;
-      if (index1 != NOT_FOUND && index2 != NOT_FOUND)
-        break;
-    }
-  }
-
   // Helper function for poset_del.
   // Function checks if there are elements between elements at idx1=A and idx2=B
   // and in relation with them. Meaning if exists C that A < C < B.
   // If such C exists function return true;
   bool somethingIsBetweenTwoElem(posetRelationsArray *relationArr,
-                                 long long const idx1, long long const idx2)
+                                 idx_t idx1, idx_t idx2)
   {
     size_t nbrOfRows = relationArr->size();
 
-    for (size_t i = 0; i < nbrOfRows; i++)
+    for (idx_t i = 0; i < nbrOfRows; i++)
     {
       if (i != idx1 && i != idx2)
       {
@@ -206,16 +188,14 @@ namespace
   // that are smaller than idx1. Meaning if exists A1,...,An that
   // A1 < elemIdx1, ..., An < elemIdx1. If such elements exist
   // function returns true.
-  bool isAnythingOnTheLeft(posetRelationsArray *relArr, long long const idx1)
+  bool isAnythingOnTheLeft(posetRelationsArray *relArr, idx_t idx1)
   {
     size_t nbrOfCol = relArr->at(idx1).size();
 
     for (size_t i = 0; i < nbrOfCol; i++)
     {
       if (relArr->at(idx1)[i] == RELATION_IM_LARGER)
-      {
         return true;
-      }
     }
     return false;
   }
@@ -225,7 +205,7 @@ namespace
   // that are larger than elemIdx2. Meaning if exists A1,...,An that
   // elemIdx1 < A1 , ..., elemIdx1 < An. If such elements exist
   // function returns true.
-  bool isAnythingOnTheRight(posetRelationsArray *relArr, long long const idx2)
+  bool isAnythingOnTheRight(posetRelationsArray *relArr, idx_t idx2)
   {
     size_t nbrOfCol = relArr->at(idx2).size();
 
@@ -251,9 +231,8 @@ namespace
   //    B < elemIdx1 < elemIdx2 < C. So allowed situations are:
   //    .... < elemIdx1 < elemIdx2 or elemIdx1 < elemIdx2 < ....
   // function returns true when all above conditions are met.
-  bool relationGoodToDelete(posetRelationsArray *relationArr,
-                            long long const idx1, long long const idx2,
-                            bool &isOnTheLeft, bool &isOnTheRight)
+  bool relationGoodToDelete(posetRelationsArray *relationArr, idx_t idx1, 
+                        idx_t idx2, bool &isOnTheLeft, bool &isOnTheRight)
   {
     if (!somethingIsBetweenTwoElem(relationArr, idx1, idx2))
     {
@@ -275,7 +254,7 @@ namespace
   // It deletes transitivity relations between elemIdx2 and elements that
   // are < elemIdx1 and have transitivity relation with elemIdx2.
   void poset_del_IsOnTheLeft(posetRelationsArray *relArr,
-                             long long const idx1, long long const idx2)
+                             idx_t idx1, idx_t idx2)
   {
     size_t nbrOfRows = relArr->at(idx2).size();
 
@@ -297,7 +276,7 @@ namespace
   // It deletes im_larger relations between elemIdx1 and elements that
   // are > elemIdx2 and have Im_larger relation with elemIdx1.
   void poset_del_IsOnTheRight(posetRelationsArray *relArr,
-                              long long const idx1, long long const idx2)
+                              idx_t idx1, idx_t idx2)
   {
     size_t nbrOfRows = relArr->at(idx2).size();
 
@@ -315,10 +294,10 @@ namespace
   }
 
   void addTransitivityRelations(posetRelationsArray *relationArr,
-                                long long const index1, long long const index2)
+                                idx_t index1, idx_t index2)
   {
     size_t nbrOfRows = relationArr->size();
-    for (index i = 0; i < nbrOfRows; i++)
+    for (idx_t i = 0; i < nbrOfRows; i++)
     {
       if (relationArr->at(i)[index1] == RELATION &&
           relationArr->at(i)[index2] == NO_RELATION)
@@ -336,9 +315,9 @@ namespace
     }
   }
 
-  #define GET_VAR_NAME(x) #x
+#define GET_VAR_NAME(x) #x
 
-  // functions for printing debbugging messages
+// functions for printing debbugging messages
   string getStrErr(char const *val)
   {
     if (val == nullptr)
@@ -352,12 +331,12 @@ namespace
   string getPairErr(string const &s1, string const &s2 = "")
   {
     if (!s2.empty())
-      return "(" + s1 + ", " + s2 + ") ";
+      return "(" + s1 + ", " + s2 + ")";
     else
       return "(" + s1 + ")";
   }
 
-  string getPosetIdErr(unsigned long id)
+  string getPosetIdErr(posetID_t id)
   {
     string str(": poset ");
     str = str + to_string(id);
@@ -381,14 +360,39 @@ namespace
     return ", " + s + " ";
   }
 
+  string relationErr(string fName, posetID_t id,
+  char const *value1, char const *value2)
+  {
+    return fName + getPosetIdErr(id) + commaElemErr("relation") + 
+        getPairErr(getStrErr(value1), getStrErr(value2));
+  }
+
 // functions that wrap the same debbugging cases
 
-void posetNotExistErr(string fName, unsigned long id)
+void posetNotExistErr(string fName, posetID_t id)
 {
   if constexpr (debug)
     {
       cerr << fName << getPosetIdErr(id) << lastExprErr();
     }
+}
+
+void elemNotExistErr(string fName, posetID_t id, char const *value)
+{
+  if constexpr (debug)
+  {
+    cerr << fName << getPosetIdErr(id) << commaElemErr() << 
+          getStrErr(value) << lastExprErr();
+  }
+}
+
+void elemExistErr(string fName, posetID_t id, char const *value)
+{
+  if constexpr (debug)
+        {
+          cerr << fName << getPosetIdErr(id) << commaElemErr() <<
+           getStrErr(value) << " already exists\n";
+        }
 }
 
 void twoValueNullErr(string fName, char const *value1, char const *value2)
@@ -405,21 +409,22 @@ void twoValueNullErr(string fName, char const *value1, char const *value2)
 void oneValueNullErr(string fName, char const *value)
 {
   if constexpr (debug)
-    {
+    {    
+      value = value;
       cerr << fName << invalidValErr(GET_VAR_NAME(value));
     }
 }
 
-void threeArgFuncNameErr(string fName, unsigned long id, char const *value1, char const *value2)
+void threeArgFuncNameErr(string fName, posetID_t id, char const *value1, char const *value2)
 {
   if constexpr (debug)
   {
-    cerr << fName << getPairErr(to_string(id),
+    cerr << fName << getPairErr(to_string(id), 
     getStrErr(value1) + ", " + getStrErr(value2)) << "\n";
-  }
+  } 
 }
 
-void twoArgFuncNameErr(string fName, unsigned long id, char const *value)
+void twoArgFuncNameErr(string fName, posetID_t id, char const *value)
 {
   if constexpr (debug)
   {
@@ -427,7 +432,7 @@ void twoArgFuncNameErr(string fName, unsigned long id, char const *value)
   }
 }
 
-void oneArgFuncNameErr(string fName, unsigned long id)
+void oneArgFuncNameErr(string fName, posetID_t id)
 {
   if constexpr (debug)
   {
@@ -435,30 +440,121 @@ void oneArgFuncNameErr(string fName, unsigned long id)
       cerr << fName << "()\n";
     else
       cerr << fName << getPairErr(to_string(id)) << "\n";
+  } 
+}
+
+void insertedErr(string fName, posetID_t id, char const *value)
+{
+  if constexpr (debug)
+      cerr << fName << getPosetIdErr(id) << commaElemErr() << 
+        getStrErr(value) << " inserted\n";
+}
+
+void containsErr(string fName, posetID_t id, size_t posetSize)
+{
+  if constexpr (debug)
+    {
+      cerr << fName << getPosetIdErr(id) << " contains " <<
+       posetSize << " element(s)\n";
+    }
+}
+
+void elemRemovedErr(string fName, posetID_t id, char const *value)
+{
+  if constexpr (debug)
+      {
+        cerr << fName << getPosetIdErr(id) << commaElemErr() << 
+          getStrErr(value) << " removed\n";
+      }
+}
+
+void isRelationAddedErr(int isAdded, string fName, posetID_t id, 
+char const *value1, char const *value2)
+{
+  if constexpr (debug)
+  {
+    if (isAdded == NO_RELATION)
+    {
+      cerr << relationErr(fName, id, value1, value2) << 
+        lastExprErr("cannot be added");
+    }
+    else if (isAdded == RELATION)
+    {
+      cerr << relationErr(fName, id, value1, value2) << 
+      lastExprErr("added");
+    }
+    
   }
 }
 
+void isRelationDeletedErr(bool isDeleted, string fName, posetID_t id, 
+char const *value1, char const *value2)
+{
+  if constexpr (debug)
+  {
+    if(isDeleted)
+    {
+      cerr << relationErr(fName, id, value1, value2) << 
+        lastExprErr("deleted");
+    }
+    else
+    {
+      cerr << relationErr(fName, id, value1, value2) << 
+        lastExprErr("cannot be deleted");
+    }
+  }
 }
 
+void relationExistsErr(bool exists, string fName, posetID_t id, 
+char const *value1, char const *value2)
+{
+  if constexpr (debug)
+  {
+    if(exists)
+    {
+      cerr << relationErr(fName, id, value1, value2) << 
+        lastExprErr("exists");
+    }
+    else
+    {
+      cerr << relationErr(fName, id, value1, value2) << 
+        lastExprErr();
+    }
+  }
+}
+
+void stateOfPosetErr(string fName, posetID_t id)
+{
+  if constexpr (debug)
+  {
+    string str;
+
+    if (fName == "poset_new")
+      str = "created";
+    else if (fName == "poset_clear")
+      str = "cleared";
+    else
+      str = "deleted";
+    
+    cerr << fName << getPosetIdErr(id) << lastExprErr(str);
+  }
+  
+}
+
+
+
+}
 
 namespace cxx {
-
-  allPosetsMap init_posets()
-  {
-    allPosetsMap initMap;
-    return initMap;
-  }
-
-allPosetsMap & getAllPosets()
+  allPosetsMap &getAllPosets()
 {
-  static allPosetsMap allPosets = init_posets();
+  static allPosetsMap allPosets; //= init_posets();
   return allPosets;
 }
-
 // tylko zeby bylo teraz, przypisuje id zawsze o 1 wiekszy od najwyzszego id.
 unsigned long poset_new(void)
 {
-  auto &allPosets = getAllPosets();
+  auto allPosets = getAllPosets();
   unsigned long id = 0;
 
   oneArgFuncNameErr(string(__func__), id);
@@ -486,25 +582,20 @@ unsigned long poset_new(void)
   poset_t *newPoset = new poset_t;
   vectorOfStrings *newVecOfPosetElem = new vectorOfStrings;
   posetRelationsArray *newRelationsArr = new posetRelationsArray;
-
   newPoset->first = newVecOfPosetElem;
   newPoset->second = newRelationsArr;
 
   allPosets.insert({id, newPoset});
-
-  if constexpr (debug)
-  {
-    cerr << __func__ << getPosetIdErr(id) << " created\n";
-  }
+  stateOfPosetErr(string(__func__), id);
 
   return id;
 }
 
-// DONE
 void poset_delete(unsigned long id)
 {
-  auto &allPosets = getAllPosets();
   oneArgFuncNameErr(string(__func__), id);
+
+  auto allPosets = getAllPosets();
 
   auto it = allPosets.find(id);
 
@@ -513,36 +604,27 @@ void poset_delete(unsigned long id)
     delete it->second->first;
     delete it->second->second;
     allPosets.erase(it);
-
-    if constexpr (debug)
-    {
-      cerr << __func__ << getPosetIdErr(id) << " deleted\n";
-    }
+    stateOfPosetErr(string(__func__), id);
   }
   else
     posetNotExistErr(string(__func__), id);
 }
 
-// DONE
 size_t poset_size(unsigned long id)
 {
-  auto &allPosets = getAllPosets();
   oneArgFuncNameErr(string(__func__), id);
+
+  auto allPosets = getAllPosets();
 
   auto it = allPosets.find(id);
   size_t sizeOfPoset = 0;
 
   if (it != allPosets.end())
   {
-    // poset_t = pair<vectorOfStrings*, posetRelationsArray*>;
-    // it->first = id, it->second = pair
     vectorOfStrings *v = it->second->first;
     sizeOfPoset = v->size();
-    if constexpr (debug)
-    {
-      cerr << __func__ << getPosetIdErr(id) << " contains " <<
-      sizeOfPoset << " element(s)\n";
-    }
+
+    containsErr(string(__func__), id, sizeOfPoset);
   }
   else
     posetNotExistErr(string(__func__), id);
@@ -552,12 +634,13 @@ size_t poset_size(unsigned long id)
 
 bool poset_insert(unsigned long id, char const *value)
 {
-  auto &allPosets = getAllPosets();
   twoArgFuncNameErr(string(__func__), id, value);
+
+  auto allPosets = getAllPosets();
 
   if (value == nullptr)
   {
-    oneValueNullErr(string(__func__), value);
+    oneValueNullErr(string(__func__), value); 
     return false;
   }
 
@@ -571,11 +654,7 @@ bool poset_insert(unsigned long id, char const *value)
     {
       if (str == value)
       {
-        if constexpr (debug)
-        {
-          cerr << __func__ << getPosetIdErr(id) << commaElemErr() << getStrErr(value) << " already exists\n";
-        }
-
+        elemExistErr(string(__func__), id, value);
         return false;
       }
     }
@@ -594,8 +673,7 @@ bool poset_insert(unsigned long id, char const *value)
     // element is in relation with itself
     p->at(p->size() - 1)[p->size() - 1] = RELATION;
 
-    if constexpr (debug)
-      cerr << __func__ << getPosetIdErr(id) << commaElemErr() << getStrErr(value) << " inserted\n";
+    insertedErr(string(__func__), id, value);
 
     return true;
   }
@@ -607,8 +685,9 @@ bool poset_insert(unsigned long id, char const *value)
 
 bool poset_remove(unsigned long id, char const *value)
 {
-  auto &allPosets = getAllPosets();
   twoArgFuncNameErr(string(__func__), id, value);
+
+  auto allPosets = getAllPosets();
 
   if (value == nullptr)
   {
@@ -648,32 +727,25 @@ bool poset_remove(unsigned long id, char const *value)
 
       relationArr->erase(relationArr->begin() + idxOfElem);
 
-      if constexpr (debug)
-      {
-        cerr << __func__ << getPosetIdErr(id) << commaElemErr() << getStrErr(value) << " removed\n";
-      }
+      elemRemovedErr(string(__func__), id, value);
 
       return true;
     }
     else
-    {
-      if constexpr (debug)
-      {
-        cerr << __func__ << getPosetIdErr(id) << commaElemErr() << getStrErr(value) << lastExprErr();
-      }
-    }
+      elemNotExistErr(string(__func__), id, value);
   }
   else
     posetNotExistErr(string(__func__), id);
-
+  
   return false;
 }
 
 bool poset_add(unsigned long id, char const *value1, char const *value2)
 {
-  auto &allPosets = getAllPosets();
   threeArgFuncNameErr(string(__func__), id, value1, value2);
 
+  auto allPosets = getAllPosets();
+  
   if (value1 == nullptr || value2 == nullptr)
   {
     twoValueNullErr(string(__func__), value1, value2);
@@ -684,28 +756,27 @@ bool poset_add(unsigned long id, char const *value1, char const *value2)
 
   if (it != allPosets.end())
   {
-    long long index1;
-    long long index2;
-    bool isRelationAdded;
+    bool foundIndex1, foundIndex2;
+    size_t index1, index2;
     vectorOfStrings *v = it->second->first;
 
-    findIndexesOfGivenValues(index1, index2, value1, value2, v);
+    checkIfElemExistInVecOfStr(v, value1, index1, foundIndex1);
+    checkIfElemExistInVecOfStr(v, value2, index2, foundIndex2);
 
     // there is no element (value1 or value2) in a set
-    if (index1 == NOT_FOUND || index2 == NOT_FOUND)
+    if (!foundIndex1 || !foundIndex2)
     {
       if constexpr (debug)
       {
         char const *val;
 
-        if (index1 == NOT_FOUND)
+        if (!foundIndex1)
           val = value1;
         else
           val = value2;
 
-        cerr << __func__ << getPosetIdErr(id) << commaElemErr() << getStrErr(val) << lastExprErr();
+        elemNotExistErr(string(__func__), id, val);
       }
-
       return false;
     }
     else
@@ -716,11 +787,7 @@ bool poset_add(unsigned long id, char const *value1, char const *value2)
       if (relationArr->at(index1)[index2] == RELATION ||
           relationArr->at(index2)[index1] == RELATION)
       {
-        if constexpr (debug)
-        {
-          cerr << __func__ << getPosetIdErr(id) << commaElemErr("relation") <<
-          getPairErr(getStrErr(value1), getStrErr(value2)) << "cannot be added\n";
-        }
+        isRelationAddedErr(NO_RELATION, string(__func__), id, value1, value2);
         return false;
       }
       else
@@ -731,9 +798,7 @@ bool poset_add(unsigned long id, char const *value1, char const *value2)
 
         // now add edges that will result from transitivity
         addTransitivityRelations(relationArr, index1, index2);
-
-        if constexpr (debug)
-          cerr << __func__ << getPosetIdErr(id) << commaElemErr("relation") << getPairErr(getStrErr(value1), getStrErr(value2)) << "added\n";
+        isRelationAddedErr(RELATION, string(__func__), id, value1, value2);
 
         return true;
       }
@@ -747,28 +812,44 @@ bool poset_add(unsigned long id, char const *value1, char const *value2)
 
 bool poset_del(unsigned long id, char const *value1, char const *value2)
 {
-  auto &allPosets = getAllPosets();
   threeArgFuncNameErr(string(__func__), id, value1, value2);
+
+  auto allPosets = getAllPosets();
 
   if (value1 == nullptr || value2 == nullptr)
   {
     twoValueNullErr(string(__func__), value1, value2);
     return false;
   }
-
+  
   auto iter = allPosets.find(id);
 
   if (iter != allPosets.end())
   {
-    long long index1, index2;
+    bool foundIdx1, foundIdx2;
+    idx_t index1, index2;
     vectorOfStrings *v = iter->second->first;
 
-    findIndexesOfGivenValues(index1, index2, value1, value2, v);
+    checkIfElemExistInVecOfStr(v, value1, index1, foundIdx1);
+    checkIfElemExistInVecOfStr(v, value2, index2, foundIdx2);
 
-    if (index1 == NOT_FOUND || index2 == NOT_FOUND)
+    if (!foundIdx1 || !foundIdx2)
+    {
+      if constexpr (debug)
+      {
+        if (!foundIdx1)
+          elemNotExistErr(string(__func__), id, value1);
+        else
+          elemNotExistErr(string(__func__), id, value2);
+      }
       return false;
+    }
     else if (index1 == index2)
+    {
+      // poset_del: poset 0, relation ("E", "G") cannot be deleted
+      isRelationDeletedErr(false, string(__func__), id, value1, value2);
       return false;
+    }
     else
     {
       posetRelationsArray *relationArr = iter->second->second;
@@ -788,9 +869,12 @@ bool poset_del(unsigned long id, char const *value1, char const *value2)
           else
             poset_del_IsOnTheRight(relationArr, index1, index2);
 
+          isRelationDeletedErr(true, string(__func__), id, value1, value2);
+
           return true;
         }
       }
+      isRelationDeletedErr(false, string(__func__), id, value1, value2);
     }
   }
   else
@@ -801,35 +885,52 @@ bool poset_del(unsigned long id, char const *value1, char const *value2)
 
 bool poset_test(unsigned long id, char const *value1, char const *value2)
 {
-  auto &allPosets = getAllPosets();
   threeArgFuncNameErr(string(__func__), id, value1, value2);
 
+  auto allPosets = getAllPosets();
+  
   if (value1 == nullptr || value2 == nullptr)
   {
     twoValueNullErr(string(__func__), value1, value2);
     return false;
   }
-
+  
   auto it = allPosets.find(id);
 
   if (it != allPosets.end())
   {
-    long long index1;
-    long long index2;
+    idx_t index1, index2;
+    bool foundIdx1, foundIdx2;
     vectorOfStrings *v = it->second->first;
 
-    findIndexesOfGivenValues(index1, index2, value1, value2, v);
+    checkIfElemExistInVecOfStr(v, value1, index1, foundIdx1);
+    checkIfElemExistInVecOfStr(v, value2, index2, foundIdx2);
 
     // there is no element (value1 or value2) in a set
-    if (index1 == NO_RELATION || index2 == NO_RELATION)
+    if (!foundIdx1 || !foundIdx2)
+    {
+      if constexpr (debug)
+      {
+        if (!foundIdx1)
+          elemNotExistErr(string(__func__), id, value1);
+        else
+          elemNotExistErr(string(__func__), id, value2);
+      }
       return false;
+    }
     else
     {
       posetRelationsArray *p = it->second->second;
 
       // if there is an edge between value1 and value2 (relation or realation transitivity)
-      if (p->at(index1)[index2] == RELATION || p->at(index1)[index2] == RELATION_TRANSITIVITY)
+      if (p->at(index1)[index2] == RELATION || 
+        p->at(index1)[index2] == RELATION_TRANSITIVITY)
+      {
+        relationExistsErr(true, string(__func__), id, value1, value2);
         return true;
+      }
+
+      relationExistsErr(false, string(__func__), id, value1, value2);
     }
   }
   else
@@ -840,9 +941,10 @@ bool poset_test(unsigned long id, char const *value1, char const *value2)
 
 void poset_clear(unsigned long id)
 {
-  auto &allPosets = getAllPosets();
   oneArgFuncNameErr(string(__func__), id);
 
+  auto allPosets = getAllPosets();
+  
   auto it = allPosets.find(id);
 
   if (it != allPosets.end())
@@ -851,272 +953,10 @@ void poset_clear(unsigned long id)
     posetRelationsArray *p = it->second->second;
     v->clear();
     p->clear();
-
-    if constexpr (debug)
-      cerr << __func__ << getPosetIdErr(id) << " cleared\n";
+    stateOfPosetErr(string(__func__), id);
   }
   else
     posetNotExistErr(string(__func__), id);;
 }
+
 }
-
-// ------------- TESTS -------------
-/*
-int getValOfTwoElemRelation(const char *val1, const char *val2, poset_t const *p)
-{
-  vectorOfStrings *v = p->first;
-  long long idx1, idx2;
-
-  findIndexesOfGivenValues(idx1, idx2, val1, val2, v);
-  if (idx1 == -1 || idx2 == -1)
-    return 2137;
-
-  return p->second->at(idx1)[idx2];
-}
-
-void TEST_poset_new_delete_insert_add()
-{
-  cout << "----- TEST_poset_new_delete_insert -----\n";
-  cout << "NEW/DELETE SECTION TEST: \n";
-
-  unsigned long id1 = poset_new();
-  unsigned long id2 = poset_new();
-  unsigned long id3 = poset_new();
-
-  assert(allPosets.size() == 3 && "allPosets.size is not equal 3, creating posets went wrong\n");
-  poset_delete(id1);
-  assert(allPosets.size() == 2 && "allPosets.size is not equal 2, deleting went wrong\n");
-
-  unsigned long id4 = poset_new();
-
-  assert(id2 == 1 && "id2 should equal 1\n");
-
-  cout << "INSERT SECTION TEST: \n";
-
-  assert(poset_insert(id1, "xd") == false && "poset id1 does not exist, but inserting was a success - WRONG\n");
-
-  size_t posetID2_size = 0;
-
-  assert(poset_insert(id2, "A") == true && "insert into id2 poset was not succesful\n");
-  assert(allPosets[id2]->first->size() == ++posetID2_size && "insert success, but num of elem not increased\n");
-  assert(poset_insert(id2, "X") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "Y") == true && "insert into id2 poset was not succesful\n");
-
-  posetID2_size += 2;
-
-  assert(poset_insert(id2, "Z") == true && "insert into id2 poset was not succesful\n");
-  assert(allPosets[id2]->first->size() == (++posetID2_size) && "insert success, but num of elem not increased");
-
-  cout << "ADD SECTION TEST: \n";
-
-  assert(poset_add(id2, "A", "B") == false);
-  assert(poset_add(id2, "A", "X") == true);
-  assert(poset_add(id2, "X", "A") == false);
-  assert(poset_add(id2, "X", "Y") == true);
-
-  cout << "TESTS PASSED\n\n";
-  cout << "printing id2=" << id2 << " poset: \n\n";
-
-  printVectorOfStrings(*allPosets[id2]->first);
-  printArrOfRelations(*allPosets[id2]->second);
-}
-
-void initPoset(long long &id1, long long &id2)
-{
-  id1 = poset_new();
-  id2 = poset_new();
-
-  assert(poset_insert(id1, "xd") == true);
-  assert(poset_insert(id1, "ALA") == true);
-
-  assert(poset_insert(id2, "A") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "B") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "C") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "X") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "Y") == true && "insert into id2 poset was not succesful\n");
-}
-
-void TEST_poset_add_remove()
-{
-  long long id1, id2;
-  initPoset(id1, id2);
-  int nbrOfPosetElem_id2 = allPosets[id2]->first->size();
-
-  assert(poset_add(id2, "A", "B") == true);
-  assert(poset_add(id2, "A", "C") == true);
-  assert(poset_add(id2, "A", "B") == false);
-
-  cout << "1) Relations added: A<B, A<C, A<B: "
-       << "\n\n";
-  printPoset(allPosets[id2]);
-
-  assert(poset_add(id2, "B", "C") == true);
-
-  cout << "2) Relations added B<C: "
-       << "\n\n";
-  printPoset(allPosets[id2]);
-
-  cout << "\nposet_remove(A):\n";
-
-  assert(poset_remove(id2, "A") == true);
-  assert(allPosets[id2]->first->size() == nbrOfPosetElem_id2 - 1 && "Removal of element of poset in vec of elems didnt work\n");
-  assert(allPosets[id2]->second->size() == nbrOfPosetElem_id2 - 1 && "Removal of whole row didnt work\n");
-  assert(allPosets[id2]->second->at(0).size() == nbrOfPosetElem_id2 - 1 && "Removal of whole column didnt work\n");
-  assert(poset_remove(id2, "A") == false);
-
-  cout << "3) printing poset after removal:\n";
-  printPoset(allPosets[id2]);
-}
-
-void DETAILED_TEST_poset_remove()
-{
-  long long id2 = poset_new();
-
-  assert(poset_insert(id2, "A") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "B") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "C") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "X") == true && "insert into id2 poset was not succesful\n");
-  assert(poset_insert(id2, "Y") == true && "insert into id2 poset was not succesful\n");
-
-  /** A < B < C, A < X, Y < X
-   *    A  B  C  X  Y
-   *   ----------------
-   * 0| 1  1  2  1 -1
-   * 1| 3  1  1 -1 -1
-   * 2| 3  3  1 -1 -1
-   * 3| 3 -1 -1  1  3
-   * 4|-1 -1 -1  1  1
-   */
-/*
-  assert(poset_add(id2, "A", "B") == true);
-  assert(poset_add(id2, "B", "C") == true);
-  assert(poset_add(id2, "A", "X") == true);
-  assert(poset_add(id2, "Y", "X") == true);
-
-  cout << "relations: A < B < C, A < X, Y < X \n\n";
-  printPoset(allPosets[id2]);
-
-  // we remove B from poset
-  assert(poset_remove(id2, "B") == true);
-  printPoset(allPosets[id2]);
-
-  // we should get
-  /** A , C, A < X, Y < X
-   *    A  C  X  Y
-   *   ----------------
-   * A| 1 -1  1 -1
-   * C|-1  1 -1 -1
-   * X| 3 -1  1  3
-   * Y|-1 -1  1  1
-   */
-  /*
-   assert(getValOfTwoElemRelation("A", "X", allPosets[id2]) == 1);
-  assert(getValOfTwoElemRelation("A", "C", allPosets[id2]) == -1);
-  assert(getValOfTwoElemRelation("C", "A", allPosets[id2]) == -1);
-
-  assert(poset_add(id2, "X", "C") == true);
-
-  printPoset(allPosets[id2]);
-  // we should get
-  /** A < X < C, Y < X < C
-   *    A  C  X  Y
-   *   ----------------
-   * A| 1  2  1 -1
-   * C| 3  1  3  3
-   * X| 3  1  1  3
-   * Y|-1  2  1  1
-   */
-
-/*
-  // C
-  assert(getValOfTwoElemRelation("C", "A", allPosets[id2]) == 3);
-  assert(getValOfTwoElemRelation("C", "X", allPosets[id2]) == 3);
-  assert(getValOfTwoElemRelation("C", "Y", allPosets[id2]) == 3);
-  // X
-  assert(getValOfTwoElemRelation("X", "A", allPosets[id2]) == 3);
-  assert(getValOfTwoElemRelation("X", "C", allPosets[id2]) == 1);
-  assert(getValOfTwoElemRelation("X", "Y", allPosets[id2]) == 3);
-}
-
-void test_peczar1()
-{
-  unsigned long p1 = poset_new();
-
-  assert(poset_size(p1) == 0);
-  assert(poset_size(p1 + 1) == 0);
-  assert(!poset_insert(p1, NULL));
-  assert(poset_insert(p1, "A"));
-  assert(poset_test(p1, "A", "A"));
-  assert(!poset_insert(p1, "A"));
-  assert(!poset_insert(p1 + 1, "B"));
-  assert(poset_size(p1) == 1);
-  assert(!poset_remove(p1 + 1, "A"));
-  assert(poset_remove(p1, "A"));
-  assert(!poset_remove(p1, "A"));
-  assert(poset_insert(p1, "B"));
-  assert(poset_insert(p1, "C"));
-  assert(poset_add(p1, "B", "C"));
-  assert(!poset_remove(p1, "A"));
-  assert(!poset_add(p1, NULL, "X"));
-  assert(!poset_del(p1, NULL, "X"));
-  assert(!poset_test(p1, NULL, "X"));
-  assert(!poset_add(p1, "X", NULL));
-  assert(!poset_del(p1, "X", NULL));
-  assert(!poset_test(p1, "X", NULL));
-  assert(!poset_add(p1, NULL, NULL));
-  assert(!poset_del(p1, NULL, NULL));
-  assert(!poset_test(p1, NULL, NULL));
-  assert(!poset_add(p1, "C", "D"));
-  assert(!poset_add(p1, "D", "C"));
-  assert(!poset_add(p1, "D", "D"));
-  assert(!poset_add(p1, "E", "D"));
-  assert(!poset_del(p1, "C", "D"));
-  assert(!poset_del(p1, "D", "C"));
-  assert(!poset_del(p1, "D", "D"));
-  assert(!poset_del(p1, "E", "D"));
-  assert(!poset_test(p1, "C", "D"));
-  assert(!poset_test(p1, "D", "C"));
-  assert(!poset_test(p1, "D", "D"));
-  assert(!poset_test(p1, "E", "D"));
-  assert(!poset_add(p1 + 1, "C", "D"));
-  assert(!poset_del(p1 + 1, "C", "D"));
-  assert(!poset_test(p1 + 1, "C", "D"));
-  poset_clear(p1);
-  poset_clear(p1 + 1);
-  assert(poset_insert(p1, "E"));
-  assert(poset_insert(p1, "F"));
-  assert(poset_insert(p1, "G"));
-  assert(poset_add(p1, "E", "F"));
-  assert(!poset_add(p1, "E", "F"));
-  assert(!poset_add(p1, "F", "E"));
-  assert(poset_test(p1, "E", "F"));
-  assert(!poset_test(p1, "F", "E"));
-  assert(poset_add(p1, "F", "G"));
-  assert(poset_test(p1, "E", "G"));
-  assert(!poset_del(p1, "E", "G"));
-  assert(poset_del(p1, "E", "F"));
-  assert(!poset_del(p1, "E", "F"));
-  assert(!poset_del(p1, "G", "F"));
-  assert(!poset_del(p1, "G", "G"));
-  assert(poset_size(p1) == 3);
-  poset_delete(p1);
-  poset_delete(p1);
-  poset_delete(p1 + 1);
-}
-*/
-
-// --------------------------------
-
-/*
-int main()
-{
-  string fName = __func__;
-  cout << "function name in string: " << fName << "\n";
-  // TEST_poset_new_delete_insert_add();
-  // TEST_poset_add_remove();
-  // DETAILED_TEST_poset_remove();
-  test_peczar1();
-
-  return 0;
-}
-*/
